processfile/src/requirements.txt
elasticsearch
netaddr
maxminddb-geolite2
